Fix code quality
